feat: respect `SourceMapDevToolPlugin` plugin for `sourceMap` option
