+++ README.md
@@ -280,7 +280,7 @@ module.exports = {
 ### `sourceMap`
 
 Type: `Boolean`
-Default: `false` (see below for details around `devtool`)
+Default: `false` (see below for details around `devtool` value and `SourceMapDevToolPlugin` plugin)
 
 **Works only with `source-map`, `inline-source-map`, `hidden-source-map` and `nosources-source-map` values for the [`devtool`](https://webpack.js.org/configuration/devtool/) option.**
 
@@ -289,8 +289,9 @@ Why?
 - `eval` wraps modules in `eval("string")` and the minimizer does not handle strings.
 - `cheap` has not column information and minimizer generate only a single line, which leave only a single mapping.
 
-The plugin respect the [`devtool`](https://webpack.js.org/configuration/devtool/).
+The plugin respect the [`devtool`](https://webpack.js.org/configuration/devtool/) and using the `SourceMapDevToolPlugin` plugin.
 Using supported `devtool` values enable source map generation.
+Using `SourceMapDevToolPlugin` with enabled the `columns` option enables source map generation.
 
 Use source maps to map error message locations to modules (this slows down the compilation).
 If you use your own `minify` function please read the `minify` section for handling source maps correctly.
