+++ package.json
@@ -40,15 +40,15 @@
     "cacache": "^13.0.1",
     "find-cache-dir": "^3.0.0",
     "jest-worker": "^24.9.0",
-    "schema-utils": "^2.4.1",
+    "schema-utils": "^2.5.0",
     "serialize-javascript": "^2.1.0",
     "source-map": "^0.6.1",
-    "terser": "^4.3.8",
+    "terser": "^4.3.9",
     "webpack-sources": "^1.4.3"
   },
   "devDependencies": {
-    "@babel/cli": "^7.6.3",
-    "@babel/core": "^7.6.3",
+    "@babel/cli": "^7.6.4",
+    "@babel/core": "^7.6.4",
     "@babel/preset-env": "^7.6.3",
     "@commitlint/cli": "^8.1.0",
     "@commitlint/config-conventional": "^8.1.0",
@@ -62,7 +62,7 @@
     "eslint": "^6.5.1",
     "eslint-config-prettier": "^6.4.0",
     "eslint-plugin-import": "^2.18.2",
-    "husky": "^3.0.8",
+    "husky": "^3.0.9",
     "jest": "^24.9.0",
     "jest-junit": "^8.0.0",
     "lint-staged": "^9.4.2",
@@ -70,7 +70,7 @@
     "npm-run-all": "^4.1.5",
     "prettier": "^1.18.2",
     "standard-version": "^7.0.0",
-    "uglify-js": "^3.6.1",
+    "uglify-js": "^3.6.3",
     "webpack": "^4.41.0"
   },
   "keywords": [
