+++ test/sourceMap-option.test.js
@@ -1,3 +1,5 @@
+import { SourceMapDevToolPlugin } from 'webpack';
+
 import TerserPlugin from '../src/index';
 
 import {
@@ -331,5 +333,151 @@ describe('sourceMap', () => {
     expect(errors).toMatchSnapshot('errors');
     expect(warnings).toMatchSnapshot('warnings');
     expect(getAssets(stats, compiler)).toMatchSnapshot('assets');
+  });
+
+  it('should match snapshot for the `SourceMapDevToolPlugin` plugin (like `source-map`)', async () => {
+    const compiler = createCompiler({
+      entry: `${__dirname}/fixtures/entry.js`,
+      devtool: false,
+      plugins: [
+        new SourceMapDevToolPlugin({
+          filename: '[file].map[query]',
+          module: true,
+          columns: true,
+        }),
+      ],
+    });
+
+    new TerserPlugin().apply(compiler);
+
+    const stats = await compile(compiler);
+
+    const errors = stats.compilation.errors.map(cleanErrorStack);
+    const warnings = stats.compilation.warnings.map(cleanErrorStack);
+
+    expect(errors).toMatchSnapshot('errors');
+    expect(warnings).toMatchSnapshot('warnings');
+    expect(getAssets(stats, compiler)).toMatchSnapshot('assets');
+  });
+
+  it('should match snapshot for the `SourceMapDevToolPlugin` plugin (like `cheap-source-map`)', async () => {
+    const compiler = createCompiler({
+      entry: `${__dirname}/fixtures/entry.js`,
+      devtool: false,
+      plugins: [
+        new SourceMapDevToolPlugin({
+          filename: '[file].map[query]',
+          module: false,
+          columns: false,
+        }),
+      ],
+    });
+
+    new TerserPlugin().apply(compiler);
+
+    const stats = await compile(compiler);
+
+    const errors = stats.compilation.errors.map(cleanErrorStack);
+    const warnings = stats.compilation.warnings.map(cleanErrorStack);
+
+    expect(errors).toMatchSnapshot('errors');
+    expect(warnings).toMatchSnapshot('warnings');
+    expect(getAssets(stats, compiler)).toMatchSnapshot('assets');
+  });
+
+  it('should match snapshot for multi compiler mode with source maps', async () => {
+    const multiCompiler = createCompiler([
+      {
+        mode: 'production',
+        devtool: 'eval',
+        bail: true,
+        cache: false,
+        entry: `${__dirname}/fixtures/entry.js`,
+        output: {
+          path: `${__dirname}/dist`,
+          filename: '[name]-1.js',
+          chunkFilename: '[id]-1.[name].js',
+        },
+        optimization: {
+          minimize: false,
+        },
+        plugins: [new TerserPlugin()],
+      },
+      {
+        mode: 'production',
+        devtool: 'source-map',
+        bail: true,
+        cache: false,
+        entry: `${__dirname}/fixtures/entry.js`,
+        output: {
+          path: `${__dirname}/dist`,
+          filename: '[name]-2.js',
+          chunkFilename: '[id]-2.[name].js',
+        },
+        optimization: {
+          minimize: false,
+        },
+        plugins: [new TerserPlugin()],
+      },
+      {
+        mode: 'production',
+        bail: true,
+        cache: false,
+        devtool: false,
+        entry: `${__dirname}/fixtures/entry.js`,
+        output: {
+          path: `${__dirname}/dist`,
+          filename: '[name]-3.js',
+          chunkFilename: '[id]-3.[name].js',
+        },
+        optimization: {
+          minimize: false,
+        },
+        plugins: [
+          new SourceMapDevToolPlugin({
+            filename: '[file].map[query]',
+            module: false,
+            columns: false,
+          }),
+          new TerserPlugin(),
+        ],
+      },
+      {
+        mode: 'production',
+        bail: true,
+        cache: false,
+        devtool: false,
+        entry: `${__dirname}/fixtures/entry.js`,
+        output: {
+          path: `${__dirname}/dist`,
+          filename: '[name]-4.js',
+          chunkFilename: '[id]-4.[name].js',
+        },
+        optimization: {
+          minimize: false,
+        },
+        plugins: [
+          new SourceMapDevToolPlugin({
+            filename: '[file].map[query]',
+            module: true,
+            columns: true,
+          }),
+          new TerserPlugin(),
+        ],
+      },
+    ]);
+
+    const multiStats = await compile(multiCompiler);
+
+    multiStats.stats.forEach((stats, index) => {
+      const errors = stats.compilation.errors.map(cleanErrorStack);
+      const warnings = stats.compilation.warnings.map(cleanErrorStack);
+
+      expect(errors).toMatchSnapshot('errors');
+      expect(warnings).toMatchSnapshot('warnings');
+      expect(getAssets(stats, multiCompiler.compilers[index])).toMatchSnapshot(
+        'assets'
+      );
+    });
   });
 });
